+++ docs/automated_model.json
@@ -1,1 +1,1 @@
-[{"isClass":false,"name":"dict","id":10,"value":{"x":81,"y":100,"z":121},"x":285,"y":25,"height":300,"width":212},{"isClass":true,"name":"Person","id":99,"value":{"age":12,"name":17},"stack_frame":false,"x":522,"y":25,"height":225,"width":198},{"isClass":false,"name":"list","id":84,"value":[32,10,90,57],"show_indexes":true,"x":745,"y":25,"height":150,"width":282},{"isClass":false,"name":"int","id":19,"value":1969,"x":1052,"y":25,"height":142,"width":212},{"isClass":false,"name":"bool","id":32,"value":true,"x":285,"y":350,"height":142,"width":212,"rowBreaker":true},{"isClass":false,"name":"str","id":43,"value":"David is cool","x":522,"y":350,"height":142,"width":212},{"isClass":false,"name":"None","id":13,"value":"None","x":759,"y":350,"height":142,"width":212},{"isClass":false,"name":"list","id":82,"value":[19,43,28,49],"x":285,"y":517,"height":130,"width":282,"rowBreaker":true},{"isClass":false,"name":"tuple","id":11,"value":[82,76],"x":592,"y":517,"height":130,"width":200},{"isClass":false,"name":"set","id":90,"value":[36,49,64],"x":817,"y":517,"height":130,"width":249}]
+[{"isClass":false,"name":"dict","id":10,"value":{"x":81,"y":100,"z":121},"x":615,"y":25,"height":300,"width":212},{"isClass":true,"name":"Person","id":99,"value":{"age":12,"name":17},"stack_frame":false,"x":842,"y":25,"height":225,"width":198},{"isClass":false,"name":"list","id":84,"value":[32,10,90,57],"show_indexes":true,"x":615,"y":340,"height":150,"width":282,"rowBreaker":true},{"isClass":false,"name":"int","id":19,"value":1969,"x":912,"y":340,"height":142,"width":212},{"isClass":false,"name":"bool","id":32,"value":true,"x":615,"y":505,"height":142,"width":212,"rowBreaker":true},{"isClass":false,"name":"str","id":43,"value":"David is cool","x":842,"y":505,"height":142,"width":212},{"isClass":false,"name":"None","id":13,"value":"None","x":1069,"y":505,"height":142,"width":212},{"isClass":false,"name":"list","id":82,"value":[19,43,28,49],"x":615,"y":662,"height":130,"width":282,"rowBreaker":true},{"isClass":false,"name":"tuple","id":11,"value":[82,76],"x":912,"y":662,"height":130,"width":200},{"isClass":false,"name":"set","id":90,"value":[36,49,64],"x":615,"y":807,"height":130,"width":249,"rowBreaker":true}]
